renaming throttle to batch in all the files
