+++ lib/index.js
@@ -26,7 +26,7 @@ var find = require('./modifiers/find.js');
 var findWhere = require('./modifiers/findWhere.js');
 var drop = require('./modifiers/drop.js');
 var take = require('./modifiers/take.js');
-var throttle = require('./modifiers/throttle.js');
+var batch = require('./modifiers/batch.js');
 
 // Consumers
 module.exports.first = v1First;
@@ -51,7 +51,7 @@ module.exports.find = find;
 module.exports.findWhere = findWhere;
 module.exports.drop = drop;
 module.exports.take = take;
-module.exports.throttle = throttle;
+module.exports.batch = batch;
 
 /**
  * Returns a stream that has been split on new lines (by default). This is a
