+++ lib/modifiers/batch.js
