+++ test/index.test.js
@@ -24,7 +24,7 @@ describe('[index]', function () {
         'findWhere',
         'drop',
         'take',
-        'throttle',
+        'batch',
         'first',
         'forEach',
         'wait'
@@ -87,7 +87,7 @@ describe('[index]', function () {
     expect(index.take).to.be.a('function')
       .and.to.have.lengthOf(1);
 
-    expect(index.throttle).to.be.a('function')
+    expect(index.batch).to.be.a('function')
       .and.to.have.lengthOf(1);
 
     expect(index.first).to.be.a('function')
